requirements.txt
asgiref==3.9.0
attrs==25.3.0
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
PyJWT==2.9.0
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==78.1.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.0
uritemplate==4.2.0
wheel==0.45.1
